SwingedgeCore/cloud/aws/scripts/spark_s3_load/requirements.txt
alpha_vantage
python-dotenv
pandas 
psycopg2-binary
asyncpg
SQLAlchemy 
sqlalchemy-timescaledb
boto3
setuptools
wheel
twine
pyspark==3.1.2
git+https://github.com/dhirs/swingedge_core.git@main#egg=SwingedgeCore







